app/mcp/pyproject.toml
[project]
name = "github-shop-mcp"
version = "0.1.0"
description = "MCP server for the shop application"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.11.0,<3.0.0",
    "opentelemetry-instrumentation-mcp",
    "opentelemetry-instrumentation-sqlalchemy",
    "opentelemetry-exporter-otlp",
    "azure-monitor-opentelemetry",
    "azure-monitor-opentelemetry-exporter",
    "opentelemetry-distro",
    # local editable packages
    "github-shop-shared",
]

[tool.uv.sources]
github-shop-shared = { path = "../shared", editable = true }

[build-system]
requires = ["uv_build>=0.9.3,<0.10.0"]
build-backend = "uv_build"